syskit: fix the order between the transformer frames and the ports
The edges where always defined as "from frame to port". However,
even with dir=none, dot will make the edge start at the left side
of an output port to go to the right side of the frame. Fix this.

Moreover, remove the "constraint=false" parameter, which improves
the final layout quite a bit.